+++ src/main/kotlin/equalexperts/number/api/models/NumberMetadataResponse.kt
@@ -3,7 +3,7 @@ package equalexperts.number.api.models
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class NumberMetadataResponse(var number : Int = 0) {
+data class NumberMetadataResponse(var number : Long = 0) {
 
     var nameOfDayOfTheWeek : DayOfTheWeekEnum  = DayOfTheWeekEnum.Unknown
