+++ src/main/kotlin/equalexperts/number/api/services/NumberService.kt
@@ -5,7 +5,7 @@ import equalexperts.number.api.functions.ExtensionFunctions.addIsEven
 import equalexperts.number.api.models.NumberMetadataResponse
 
 class NumberService : NumberServiceInterface {
-    override fun getNumberMetadata(number: Int): NumberMetadataResponse {
+    override fun getNumberMetadata(number: Long): NumberMetadataResponse {
         return  NumberMetadataResponse()
             .addDayOfTheWeek(number)
             .addIsEven(number)
