+++ src/main/kotlin/equalexperts/number/api/services/NumberServiceInterface.kt
@@ -3,5 +3,5 @@ package equalexperts.number.api.services
 import equalexperts.number.api.models.NumberMetadataResponse
 
 interface NumberServiceInterface {
-    fun getNumberMetadata(number : Int) : NumberMetadataResponse
+    fun getNumberMetadata(number : Long) : NumberMetadataResponse
 }
